add  step-2 pytest
